refactorL simplify log (disabled by level in module)
